+++ src/rust/rust-proto-new/tests/test_utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod strategies;
